py_requirements.txt
colorama==0.4.1
jinja2==3.0.3
netaddr==0.7.19
pyyaml==5.4.1
ruamel.yaml==0.17.17
